# Goflow parse