--- a/go.mod
+++ b/go.mod
@@ -9,9 +9,7 @@ require (
 	github.com/aws/aws-sdk-go v1.38.31
 	github.com/bmatcuk/doublestar/v2 v2.0.4
 	github.com/cenkalti/backoff/v4 v4.1.1
-	github.com/cloudflare/goflow/v3 v3.4.2
 	github.com/elastic/go-elasticsearch/v7 v7.13.0
-	github.com/fatih/structs v1.1.0
 	github.com/golang/protobuf v1.5.2
 	github.com/hashicorp/go-uuid v1.0.2
 	github.com/jpillora/backoff v1.0.0
@@ -20,6 +18,7 @@ require (
 	github.com/mitchellh/mapstructure v1.4.1
 	github.com/observiq/ctimefmt v1.0.0
 	github.com/observiq/go-syslog/v3 v3.0.2
+	github.com/observiq/goflow/v3 v3.4.4
 	github.com/observiq/nanojack v0.0.0-20201106172433-343928847ebc
 	github.com/spf13/cobra v1.1.3
 	github.com/stretchr/testify v1.7.0