--- a/operator/builtin/input/goflow/goflow.go
+++ b/operator/builtin/input/goflow/goflow.go
@@ -7,9 +7,9 @@ import (
 	"sync"
 	"time"
 
-	flowmessage "github.com/cloudflare/goflow/v3/pb"
-	"github.com/cloudflare/goflow/v3/utils"
 	"github.com/jpillora/backoff"
+	flowmessage "github.com/observiq/goflow/v3/pb"
+	"github.com/observiq/goflow/v3/utils"
 	"github.com/observiq/stanza/operator"
 	"github.com/observiq/stanza/operator/helper"
 	"go.uber.org/zap"