--- a/operator/builtin/input/goflow/parse.go
+++ b/operator/builtin/input/goflow/parse.go
@@ -17,7 +17,9 @@ func Parse(message *flowmessage.FlowMessage) (map[string]interface{}, time.Time,
 	timestamp := time.Unix(int64(message.TimeReceived), 0)
 
 	if t := message.Type.String(); t != "" {
-		m["type"] = t
+		if t != "FLOWUNKNOWN" {
+			m["type"] = t
+		}
 	}
 
 	if message.SequenceNum > 0 {