… go.sum
… operator/builtin/input/goflow/parse_protocol_test.go