## Description of Changes

- Switched to observiq fork of Goflow
  - This fork sets a boolean value if "FlowDirection / BiFlowDirection" is received from the netflow source. This is because 0 is a valid value, and the json parser was filtering out zeros due to the omit empty tag. We want to skip this field if it was not sent from the netflow client.
- Replaced `structs.Map()` call (converting goflow message to map[string]interface{} in favor of manual parsing.
  - This resolves an issue where proto field is missing when 0 (0 == ICMP)
- Moved proto_name parsing from parse.go to parse_protocol.go 
- Parse() accepts pointer value to prevent copying mutex lock

This new behavior means test cases need to set the following to their zero values:
```
				"proto":          0,
				"proto_name":     "HOPOPT",
				"inif":           0,
				"outif":          0,
```

Proto, InIf, OutIf are always present (from what I have seen in the field). If they are not present, the user will see zero values and the `proto_name` "HOPOPT". It is possible a future version of the Goflow package will allow us to detect if the fields were sent from the client.

I observed significant performance increase

Master:
```
go test ./operator/builtin/input/goflow -bench=. -benchtime=10s

goos: linux
goarch: amd64
pkg: github.com/observiq/stanza/operator/builtin/input/goflow
cpu: AMD Ryzen 9 3950X 16-Core Processor       
     
BenchmarkParse-32    	  102042	    118210 ns/op
BenchmarkParse-32    	   96640	    117744 ns/op
BenchmarkParse-32    	  101474	    116649 ns/op

PASS

ok  	github.com/observiq/stanza/operator/builtin/input/goflow	13.263s
```

This PR:
```
go test ./operator/builtin/input/goflow -bench=. -benchtime=10s

goos: linux
goarch: amd64
pkg: github.com/observiq/stanza/operator/builtin/input/goflow
cpu: AMD Ryzen 9 3950X 16-Core Processor            

BenchmarkParse-32    	  971654	     12851 ns/op
BenchmarkParse-32    	 1000000	     12928 ns/op
BenchmarkParse-32    	 1000000	     12941 ns/op

PASS

ok  	github.com/observiq/stanza/operator/builtin/input/goflow	12.621s
```

## **Please check that the PR fulfills these requirements**
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] Add a changelog entry (for non-trivial bug fixes / features)
- [x] CI passes