@djaglowski 2021-07-15
| Log Files | Logs / Second | CPU Avg (%) | CPU Avg Δ (%) | Memory Avg (MB) | Memory Avg Δ (MB) |
| --- | --- | --- | --- | --- | --- |
|1|1000|1.6551741|+0.22406816|128.74258|-1.8078003|
|1|5000|5.0173345|-0.31027317|136.39964|-0.2883911|
|1|10000|9.982939|+0.051854134|146.44894|-3.0464783|
|1|50000|48.277054|-1.8439255|171.36153|-7.477646|
|1|100000|95.431564|-1.9628143|226.53435|-1.8065796|
|10|100|2.1553032|+0.27589774|134.56923|-0.55766296|
|10|500|5.7932997|-0.3792057|140.03596|-0.6356354|
|10|1000|11.414141|+0.017298698|147.43669|-0.8504791|
|10|5000|53.864235|-0.28815842|172.82785|-5.1876373|
|10|10000|103.494804|-1.9001007|214.06627|-8.561554|