@djaglowski 2021-07-15
| Log Files | Logs / Second | CPU Avg (%) | CPU Avg Δ (%) | Memory Avg (MB) | Memory Avg Δ (MB) |
| --- | --- | --- | --- | --- | --- |
|1|1000|1.4310849|-2.1100044e-05|128.98856|-1.5618286|
|1|5000|5.172461|-0.1551466|137.84093|+1.1528931|
|1|10000|10.5349045|+0.60381985|146.47858|-3.0168457|
|1|50000|49.379807|-0.7411728|174.90907|-3.9300995|
|1|100000|95.82563|-1.5687485|233.09267|+4.7517395|
|10|100|1.7931755|-0.08623004|133.53085|-1.5960388|
|10|500|6.672552|+0.50004673|137.78569|-2.88591|
|10|1000|12.362235|+0.96539307|146.78812|-1.499054|
|10|5000|58.222443|+4.0700493|177.26292|-0.7525635|
|10|10000|105.724365|+0.32946014|225.91136|+3.2835388|